requirements.txt
aiohttp==3.7.2
aioresponses==0.7.1
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
certstream==1.11
cffi==1.14.3
chardet==3.0.4
configparser==5.0.1
cryptography==3.1.1
decorator==4.4.2
dnspython==2.0.0
elasticsearch==7.9.1
httpretty==1.0.2
idna==2.10
lxml==4.5.1
multidict==5.0.2
pycparser==2.20
pycrtsh==0.3.2
pysecuritytrails==0.1.3
python-dateutil==2.8.1
python-telegram-bot==13.0
python-Wappalyzer==0.3.1
pytz==2020.1
requests==2.24.0
schedule==0.6.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
termcolor==1.1.0
tld==0.12.2
tornado==6.0.4
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
websocket-client==0.57.0
yarl==1.6.3
